+++ packages/ic_sis_provider/src/service/sis_prepare_login.rs
@@ -10,14 +10,15 @@ struct PrepareLoginOkResponse {
 
 #[update]
 fn sis_prepare_login(address: String) -> Result<PrepareLoginOkResponse, String> {
-    // Create a SuiAddress from the string. This validates the address.
     let address = SuiAddress::new(&address)?;
 
     match ic_sis::login::prepare_login(&address) {
-        Ok(m) => Ok(PrepareLoginOkResponse {
-            sis_message: m.0.into(),
-            nonce: m.1,
-        }),
+        Ok((message, nonce)) => {
+            Ok(PrepareLoginOkResponse {
+                sis_message: message.to_human_readable(), // Use human-readable format
+                nonce,
+            })
+        }
         Err(e) => Err(e.into()),
     }
 }